refactor(game): move TextureVideo into a separate file
